rootlender-repayment-service/requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
psycopg2-binary
pydantic
python-dotenv
httpx
